hey_spider_robot/requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
opencv-python==4.8.1.78
ultralytics==8.0.196
adafruit-circuitpython-pca9685==3.4.15
adafruit-circuitpython-servokit==1.3.13
adafruit-circuitpython-ssd1306==2.12.14
Pillow==10.1.0
SpeechRecognition==3.10.0
pyaudio==0.2.11
openai==1.3.0
numpy==1.24.3
python-socketio==5.8.0
RPi.GPIO==0.7.1
gpiozero==1.6.2